+++ common/org.eclipse.escet.common.raildiagrams/src/org/eclipse/escet/common/raildiagrams/railroad/DiagramElement.java
@@ -129,16 +129,16 @@ public abstract class DiagramElement {
      * @param xOffset Horizontal offset of the element in the picture.
      * @param yOffset Vertical offset of the element in the picture.
      */
-    public void dumpElementBox(String coordType, double xOffset, double yOffset) {
+    public void dumpElementBox(String coordType, int xOffset, int yOffset) {
         dbg("%s coordinates:", coordType);
         idbg();
-        dbg("%s-%s: x[%.1f--%.1f], y[%.1f--%.1f], connectTop=%.1f", kindName, id, xOffset + solver.getVarValue(left),
+        dbg("%s-%s: x[%d--%d], y[%d--%d], connectTop=%d", kindName, id, xOffset + solver.getVarValue(left),
                 xOffset + solver.getVarValue(right), yOffset + solver.getVarValue(top),
                 yOffset + solver.getVarValue(bottom), yOffset + solver.getVarValue(connectTop));
         idbg();
         for (ProxyDiagramElement proxy: childDiagramElements) {
             DiagramElement child = proxy.child;
-            dbg("%s-%s: x[%.1f--%.1f], y[%.1f--%.1f], connectTop=%.1f", child.kindName, child.id,
+            dbg("%s-%s: x[%d--%d], y[%d--%d], connectTop=%d", child.kindName, child.id,
                     xOffset + solver.getVarValue(proxy.left), xOffset + solver.getVarValue(proxy.right),
                     yOffset + solver.getVarValue(proxy.top), yOffset + solver.getVarValue(proxy.bottom),
                     yOffset + solver.getVarValue(proxy.connectTop));
@@ -159,7 +159,7 @@ public abstract class DiagramElement {
      * @param outputTarget Diagram to write.
      * @param dumpAbsCoords Whether to dump the absolute coordinates of the elements for debugging.
      */
-    public void paint(double left, double top, OutputTarget outputTarget, boolean dumpAbsCoords) {
+    public void paint(int left, int top, OutputTarget outputTarget, boolean dumpAbsCoords) {
         outputTarget.addDiagramElement(left, top, solver, this);
 
         if (dumpAbsCoords) {
@@ -174,8 +174,8 @@ public abstract class DiagramElement {
         }
 
         for (ProxyDiagramElement childElement: childDiagramElements) {
-            double childLeft = left + solver.getVarValue(childElement.left);
-            double childTop = top + solver.getVarValue(childElement.top);
+            int childLeft = left + solver.getVarValue(childElement.left);
+            int childTop = top + solver.getVarValue(childElement.top);
             childElement.paint(childLeft, childTop, outputTarget, dumpAbsCoords);
         }
     }
